Consolidate type files
