--- src/v2/util/types.d.ts
@@ -1,6 +1,0 @@
-export type Workers = {
-	hack: string;
-	grow: string;
-	weaken: string;
-	all: string[];
-};
